+++ VideoApp/VideoApp/Stores/AppSettings/AppSettingsStore.swift
@@ -19,6 +19,7 @@ import Foundation
 protocol AppSettingsStoreWriting: LaunchStore {
     var environment: Environment { get set }
     var videoCodec: VideoCodec { get set }
+    var videoSize: VideoSize { get set }
     var userIdentity: String { get set }
     var isTURNMediaRelayOn: Bool { get set }
     var areInsightsEnabled: Bool { get set }
@@ -39,6 +40,7 @@ protocol AppSettingsStoreWriting: LaunchStore {
 class AppSettingsStore: AppSettingsStoreWriting {
     @Storage(key: makeKey("environment"), defaultValue: Environment.production) var environment: Environment
     @Storage(key: makeKey("videoCodec"), defaultValue: defaultVideoCodec) var videoCodec: VideoCodec
+    @Storage(key: makeKey("videoSize"), defaultValue: .vga) var videoSize: VideoSize
     @Storage(key: makeKey("userIdentity"), defaultValue: "") var userIdentity: String
     @Storage(key: makeKey("isTURNMediaRelayOn"), defaultValue: false) var isTURNMediaRelayOn: Bool
     @Storage(key: makeKey("areInsightsEnabled"), defaultValue: true) var areInsightsEnabled: Bool
@@ -65,7 +67,7 @@ class AppSettingsStore: AppSettingsStoreWriting {
     private let userDefaults: UserDefaultsProtocol
     private static var defaultVideoCodec: VideoCodec {
         switch appInfoStore.appInfo.target {
-        case .videoInternal: return .vp8SimulcastVGA
+        case .videoInternal: return .auto
         case .videoCommunity: return .vp8
         }
     }