+++ VideoApp/VideoApp/Stores/RemoteConfig/RemoteConfigStore.swift
@@ -48,14 +48,8 @@ class RemoteConfigStore: RemoteConfigStoreWriting {
         }
         set {
             guard newValue != appSettingsStore.remoteRoomType else { return }
-            
-            switch newValue {
-            case .group, .groupSmall, .unknown:
-                appSettingsStore.videoCodec = .vp8SimulcastVGA
-            case .peerToPeer, .go:
-                appSettingsStore.videoCodec = .vp8
-            }
 
+            appSettingsStore.videoCodec = .auto
             appSettingsStore.remoteRoomType = newValue
         }
     }